upgrade: pipe側にパーミッション処理を追加
