+++ command/src/pipe_process.c
@@ -6,7 +6,7 @@
 /*   By: tkirihar <tkirihar@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/10 21:07:42 by tkirihar          #+#    #+#             */
-/*   Updated: 2022/02/23 22:48:44 by tkirihar         ###   ########.fr       */
+/*   Updated: 2022/03/03 16:35:40 by tkirihar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -151,9 +151,15 @@ void	exec_cmd(t_exec_attr *ea, t_pipe_attr *pa)
 				cmd_path = find_path(pa->current_cmd->cmd, ea);
 				if (cmd_path == NULL)
 				{
-					printf("%s: command not found\n", pa->current_cmd->cmd); // 標準エラー出力にする
+					ft_put_error("command not found", pa->current_cmd->cmd);
 					exit(127);
 				}
+			}
+			// cmd_pathがディレクトリか確認する処理
+			if (is_dir(cmd_path))
+			{
+				ft_put_error("is a directory", pa->current_cmd->cmd);
+				exit(126);
 			}
 		}
 		if (has_redirect_file(pa->current_cmd))
